Benchmark/requirements.txt
absl-py==2.0.0
accelerate==0.31.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
auditwheel==5.4.0
Babel==2.13.1
beautifulsoup4==4.12.2
bertviz==1.4.0
bleach==6.1.0
blinker==1.6.3
blis==0.7.11
boto3==1.28.77
botocore==1.31.77
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.1
click==8.1.7
cloudpathlib==0.16.0
comm==0.1.4
confection==0.1.3
contexttimer==0.3.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
datasets==2.16.1
debugpy==1.8.0
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.16.0
dill==0.3.7
distlib==0.3.7
distro==1.9.0
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
ez_setup==0.9
fairscale==0.4.4
fastchat==0.1.0
fastjsonschema==2.18.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
identify==2.5.31
idna==3.4
imageio==2.31.6
importlib-metadata==6.8.0
iopath==0.1.10
ipykernel==6.26.0
ipython==8.17.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kaggle==1.5.16
keras==2.15.0
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nodeenv==1.8.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.35.7
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
openpyxl==3.1.5
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandoc==2.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
Pillow==10.0.1
platformdirs==3.11.0
plotly==5.18.0
plumbum==1.8.2
ply==3.11
portalocker==2.8.2
pre-commit==3.5.0
preshed==3.0.9
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pyelftools==0.30
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.28.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminado==0.17.1
text-unidecode==1.3
thinc==8.2.1
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.43.0
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==2.0.7
validators==0.22.0
virtualenv==20.24.6
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.9
weasel==0.3.3
webcolors==1.13
webdataset==0.2.71
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
